# DataBase/apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: DataBase Gateway
  version: 0.1.0
  description: Thin API facade for the DataBase ecosystem.
servers:
  - url: http://127.0.0.1:18090
components:
  securitySchemes:
    DataBaseApiKey:
      type: apiKey
      in: header
      name: X-DataBase-Api-Key
      description: Optional. Required only when DATABASE_GATEWAY_AUTH_REQUIRED=true.
  schemas:
    RequestId:
      type: string
      description: Per-request correlation id returned in X-Request-Id.
...
    ErrorResponse:
      type: object
      required:
        - ok
        - error
        - message
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: string
        message:
          type: string
        requestId:
          $ref: "#/components/schemas/RequestId"
...
    OpenListFileObject:
      type: object
      properties:
        name:
          type: string
        size:
          type: number
        is_dir:
          type: boolean
        modified:
          type: string
        created:
          type: string
        sign:
          type: string
        thumb:
          type: string
        type:
          type: number
        hashinfo:
          type: string
        hash_info:
          type: object
          propertyNames:
            type: string
          additionalProperties:
            type: string
      required:
        - name
        - size
        - is_dir
        - modified
        - created
      additionalProperties: {}
    OpenListStoragesResponse:
      type: object
      properties:
        count:
          type: number
        storages:
          type: array
          items:
            type: object
            properties:
              id:
                type: number
              mount_path:
                type: string
              order:
                type: number
              driver:
                type: string
              cache_expiration:
                type: number
              status:
                type: string
              addition:
                type: string
              remark:
                type: string
              disabled:
                type: boolean
            required:
              - id
              - mount_path
            additionalProperties: {}
        requestId:
          type: string
      required:
        - count
        - storages
        - requestId
      additionalProperties: false
    OpenListStorageResponse:
      type: object
      properties:
        storage:
          type: object
          properties:
            id:
              type: number
            mount_path:
              type: string
            order:
              type: number
            driver:
              type: string
            cache_expiration:
              type: number
            status:
              type: string
            addition:
              type: string
            remark:
              type: string
            disabled:
              type: boolean
          required:
            - id
            - mount_path
          additionalProperties: {}
        requestId:
          type: string
      required:
        - storage
        - requestId
      additionalProperties: false
    OpenListFsListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              size:
                type: number
              is_dir:
                type: boolean
              modified:
                type: string
              created:
                type: string
              sign:
                type: string
              thumb:
                type: string
              type:
                type: number
              hashinfo:
                type: string
              hash_info:
                type: object
                propertyNames:
                  type: string
                additionalProperties:
                  type: string
            required:
              - name
              - size
              - is_dir
              - modified
              - created
            additionalProperties: {}
        total:
          type: number
        readme:
          type: string
        header:
          type: string
        write:
          type: boolean
        write_content_bypass:
          type: boolean
        provider:
          type: string
        direct_upload_tools:
          type: array
          items:
            type: string
        requestId:
          type: string
      required:
        - content
        - total
        - requestId
      additionalProperties: false
    OpenListFsGetResponse:
      type: object
      properties:
        item:
          type: object
          properties:
            name:
              type: string
            size:
              type: number
            is_dir:
              type: boolean
            modified:
              type: string
            created:
              type: string
            sign:
              type: string
            thumb:
              type: string
            type:
              type: number
            hashinfo:
              type: string
            hash_info:
              type: object
              propertyNames:
                type: string
              additionalProperties:
                type: string
          required:
            - name
            - size
            - is_dir
            - modified
            - created
          additionalProperties: {}
        requestId:
          type: string
      required:
        - item
        - requestId
      additionalProperties: false
    OpenListMount:
      type: object
      properties:
        id:
          type: string
        mountPath:
          type: string
        driver:
          anyOf:
            - type: string
            - type: "null"
        remark:
          anyOf:
            - type: string
            - type: "null"
        openlistStatus:
          anyOf:
            - type: string
            - type: "null"
        disabled:
          type: boolean
        source:
          type: string
        metadata:
          type: object
          propertyNames:
            type: string
          additionalProperties: {}
        lastSyncedAt:
          anyOf:
            - type: string
            - type: "null"
        createdAt:
          anyOf:
            - type: string
            - type: "null"
        updatedAt:
          anyOf:
            - type: string
            - type: "null"
      required:
        - id
        - mountPath
        - driver
        - remark
        - openlistStatus
        - disabled
        - source
        - metadata
        - lastSyncedAt
        - createdAt
        - updatedAt
      additionalProperties: false
    OpenListTarget:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        purpose:
          type: string
        displayName:
          type: string
        mountPath:
          type: string
        remoteDir:
          type: string
        localCachePath:
          anyOf:
            - type: string
            - type: "null"
        status:
          type: string
        metadata:
          type: object
          propertyNames:
            type: string
          additionalProperties: {}
        createdAt:
          anyOf:
            - type: string
            - type: "null"
        updatedAt:
          anyOf:
            - type: string
            - type: "null"
      required:
        - id
        - provider
        - purpose
        - displayName
        - mountPath
        - remoteDir
        - localCachePath
        - status
        - metadata
        - createdAt
        - updatedAt
      additionalProperties: false
    OpenListMountsResponse:
      type: object
      properties:
        count:
          type: number
        mounts:
          type: array
          items:
            $ref: "#/components/schemas/OpenListMount"
        requestId:
          type: string
      required:
        - count
        - mounts
        - requestId
      additionalProperties: false
    OpenListTargetsResponse:
      type: object
      properties:
        count:
          type: number
        targets:
          type: array
          items:
            $ref: "#/components/schemas/OpenListTarget"
        requestId:
          type: string
      required:
        - count
        - targets
        - requestId
      additionalProperties: false
    OpenListTargetResponse:
      type: object
      properties:
        target:
          $ref: "#/components/schemas/OpenListTarget"
        requestId:
          type: string
      required:
        - target
        - requestId
      additionalProperties: false
    OpenListTargetFsListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OpenListFileObject"
        total:
          type: number
        readme:
          type: string
        header:
          type: string
        write:
          type: boolean
        write_content_bypass:
          type: boolean
        provider:
          type: string
        direct_upload_tools:
          type: array
          items:
            type: string
        requestId:
          type: string
        target:
          $ref: "#/components/schemas/OpenListTarget"
      required:
        - content
        - total
        - requestId
        - target
      additionalProperties: false
...
    OpenListTargetFsGetResponse:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/OpenListFileObject"
        requestId:
          type: string
        target:
          $ref: "#/components/schemas/OpenListTarget"
      required:
        - item
        - requestId
        - target
      additionalProperties: false
...
paths:
...
  /openlist/storages:
    get:
      operationId: listOpenListStorages
      summary: List OpenList storage mounts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 200
      responses:
        "200":
          description: OpenList storage mounts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListStoragesResponse"
        "401":
          description: Missing or invalid API key.
        "503":
          description: OpenList client is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /openlist/storages/{id}:
    get:
      operationId: getOpenListStorage
      summary: Get one OpenList storage mount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: One OpenList storage mount.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListStorageResponse"
        "400":
          description: Invalid storage id.
        "401":
          description: Missing or invalid API key.
        "503":
          description: OpenList client is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /openlist/fs/list:
    post:
      operationId: listOpenListFiles
      summary: List files through OpenList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - path
              properties:
                path:
                  type: string
                password:
                  type: string
                page:
                  type: integer
                  minimum: 1
                per_page:
                  type: integer
                  minimum: 1
                  maximum: 1000
                refresh:
                  type: boolean
                subPath:
                  type: string
      responses:
        "200":
          description: File list from OpenList.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListFsListResponse"
        "401":
          description: Missing or invalid API key.
        "503":
          description: OpenList client is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /openlist/fs/get:
    post:
      operationId: getOpenListFile
      summary: Get one file object through OpenList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - path
              properties:
                path:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: One file object from OpenList.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListFsGetResponse"
        "401":
          description: Missing or invalid API key.
        "503":
          description: OpenList client is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /openlist/mounts:
    get:
      operationId: listOpenListMounts
      summary: List DataBase-owned OpenList mount topology
      description: Reads mount topology from DataBase. OpenList is the access projection; the mounted backend is the file truth.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: DataBase-owned OpenList mount records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListMountsResponse"
        "401":
          description: Missing or invalid API key.
  /openlist/targets:
    get:
      operationId: listOpenListTargets
      summary: List DataBase-owned OpenList file targets
      description: Returns canonical target ids so callers do not hard-code OpenList remote paths.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: status
          in: query
          schema:
            type: string
            default: active
        - name: purpose
          in: query
          schema:
            type: string
      responses:
        "200":
          description: DataBase-owned OpenList targets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListTargetsResponse"
        "401":
          description: Missing or invalid API key.
  /openlist/targets/{id}:
    get:
      operationId: getOpenListTarget
      summary: Get one DataBase-owned OpenList file target
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: One DataBase-owned OpenList target.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListTargetResponse"
        "401":
          description: Missing or invalid API key.
        "404":
          description: OpenList target not found.
  /openlist/targets/{id}/list:
    post:
      operationId: listOpenListTargetFiles
      summary: List files for a DataBase-owned OpenList target
      description: Resolves target id from MySQL, then calls OpenList with the registered remoteDir.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                page:
                  type: integer
                  minimum: 1
                per_page:
                  type: integer
                  minimum: 1
                  maximum: 1000
                refresh:
                  type: boolean
                subPath:
                  type: string
      responses:
        "200":
          description: File list for the registered target.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListTargetFsListResponse"
        "401":
          description: Missing or invalid API key.
        "404":
          description: OpenList target not found.
        "503":
          description: OpenList client is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /openlist/targets/{id}/get:
    post:
      operationId: getOpenListTargetFile
      summary: Get one file for a DataBase-owned OpenList target
      description: Resolves the registered target id and accepts a stable bookId/object id so callers do not pass raw OpenList paths.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: string
                path:
                  type: string
                password:
                  type: string
              anyOf:
                - required:
                    - bookId
                - required:
                    - path
      responses:
        "200":
          description: One file for the registered target.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenListTargetFsGetResponse"
        "400":
          description: Invalid target file request.
        "401":
          description: Missing or invalid API key.
        "404":
          description: OpenList target or file not found.
        "503":
          description: OpenList client is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...